server/pdf_service/requirements.txt
fastapi
python-multipart
uvicorn
pdfminer.six 